Extract format and lint rules to their own modules
When all the rules are in one module,
and each rule depends on the module,
any change to the module causes the rule to be rebuilt.

For something like Haskell, that's not ideal.
A change to the module effectively means the code has to be re-compiled
as all the Haskell rules are in one file.

Splitting the rules to their own modules means rules can depend on
just the module itself.
It's not the full granularity of depending on the rule alone,
but it should be enough without trying to figure out that situation.
